COMP7940-Group1-Project/requirements.txt
telegram 
python-telegram-bot>=13.1
configparser
pyrebase4
Flask==2.1.0
gunicorn==20.1.0